Fix The sort mithods.

+++ StarkFamily/Controllers/PersonsController.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Dynamic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
@@ -19,12 +20,49 @@ namespace StarkFamily.Controllers
             _context = context;
         }
 
+       
         // GET: Persons
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(string sortOrder)
         {
-            var familyContext = _context.Personos.Include(p => p.Father);
-            return View(await familyContext.ToListAsync());
+            ViewData["NameSortParm"] = String.IsNullOrEmpty(sortOrder) ? "name_desc" : "";
+            ViewData["DateSortParm"] = sortOrder == "Date" ? "date_desc" : "Date";
+
+            //var persons = from s in _context.Personos select s;
+            IQueryable<Person> fathers = _context.Personos.Include(p => p.Father);
+
+
+            switch (sortOrder)
+            {
+                case "name_desc":
+                    fathers = fathers.OrderByDescending(s => s.FirstName);
+                    break;
+                case "Date":
+                    fathers = fathers.OrderBy(s => s.BirthDate);
+                    break;
+                case "date_desc":
+                    fathers = fathers.OrderByDescending(s => s.BirthDate);
+                    break;
+                default:
+                    fathers = fathers.OrderBy(s => s.FirstName);
+                    break;
+            }
+            //var _persons = await persons.Where(p => p.FatherId == null).AsNoTracking().ToListAsync();
+            //_persons.AddRange(await fathers.Where(p => p.FatherId != null).OrderBy(s => s.FirstName).ToListAsync());
+
+            //_persons.OrderBy(s => s.FirstName);
+
+            //return View(_persons);
+            //return View(await persons.AsNoTracking().ToListAsync());
+            return View(await fathers.ToListAsync());
+
+
+
+
+
         }
+
+        // GET: Father
+        
 
         // GET: Persons/Details/5
         public async Task<IActionResult> Details(int? id)
@@ -42,7 +80,12 @@ namespace StarkFamily.Controllers
                 return NotFound();
             }
 
-            var childe = await _context.Personos.Where(p => p.FatherId == id.Value).ToListAsync(); 
+            var childe = await _context.Personos.Where(p => p.FatherId == id.Value).OrderBy(s => s.FirstName).ToListAsync();
+
+            if (childe == null)
+            {
+                return NotFound();
+            }
 
             return View(person);
         }
@@ -50,7 +93,7 @@ namespace StarkFamily.Controllers
         // GET: Persons/Create
         public IActionResult Create()
         {
-            ViewData["FatherId"] = new SelectList(_context.Personos, "Id", "Id");
+            ViewData["FatherId"] = new SelectList(_context.Personos, "Id", "FirstName");
             return View();
         }
 
@@ -59,15 +102,27 @@ namespace StarkFamily.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,FirstName,LastName,BirthDate,DeathDate,FatherId")] Person person)
+        public async Task<IActionResult> Create([Bind("FirstName,LastName,BirthDate,DeathDate,FatherId")] Person person)
         {
-            if (ModelState.IsValid)
+            try
             {
-                _context.Add(person);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+
+                if (ModelState.IsValid)
+                {
+                    _context.Add(person);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
             }
-            ViewData["FatherId"] = new SelectList(_context.Personos, "Id", "Id", person.FatherId);
+            catch (DbUpdateException /* ex */)
+            {
+                //Log the error (uncomment ex variable name and write a log.
+                ModelState.AddModelError("", "Unable to save changes. " +
+                    "Try again, and if the problem persists " +
+                    "see your system administrator.");
+            }
+
+            ViewData["FatherId"] = new SelectList(_context.Personos, "FatherId", "FatherId", person.FirstName);
             return View(person);
         }
 
@@ -84,48 +139,46 @@ namespace StarkFamily.Controllers
             {
                 return NotFound();
             }
-            ViewData["FatherId"] = new SelectList(_context.Personos, "Id", "Id", person.FatherId);
+            ViewData["FatherId"] = new SelectList(_context.Personos, "Id", "FirstName");
             return View(person);
         }
 
         // POST: Persons/Edit/5
         // To protect from overposting attacks, enable the specific properties you want to bind to.
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
+        [HttpPost, ActionName("Edit")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,FirstName,LastName,BirthDate,DeathDate,FatherId")] Person person)
+        public async Task<IActionResult> EditPost(int? id)
         {
-            if (id != person.Id)
+            if (id == null)
             {
                 return NotFound();
             }
 
-            if (ModelState.IsValid)
+            var personToUpdate = await _context.Personos.FirstOrDefaultAsync(s => s.Id == id);
+            if (await TryUpdateModelAsync<Person>(
+                    personToUpdate,
+                    "",
+                s => s.FirstName, s => s.LastName, s => s.BirthDate, s => s.DeathDate, s => s.FatherId))
             {
                 try
                 {
-                    _context.Update(person);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (DbUpdateException /* ex */)
                 {
-                    if (!PersonExists(person.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    ModelState.AddModelError("", "Unable to save changes. " +
+                                            "Try again, and if the problem persists, " +
+                                            "see your system administrator.");
                 }
-                return RedirectToAction(nameof(Index));
             }
-            ViewData["FatherId"] = new SelectList(_context.Personos, "Id", "Id", person.FatherId);
-            return View(person);
+            ViewData["FatherId"] = new SelectList(_context.Personos, "Id", "Id");
+            return View(personToUpdate);
         }
 
         // GET: Persons/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -133,11 +186,19 @@ namespace StarkFamily.Controllers
             }
 
             var person = await _context.Personos
+                .AsNoTracking()
                 .Include(p => p.Father)
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (person == null)
             {
                 return NotFound();
+            }
+
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                    "Delete failed. Try again, and if the problem persists " +
+                    "see your system administrator.";
             }
 
             return View(person);
@@ -149,9 +210,22 @@ namespace StarkFamily.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var person = await _context.Personos.FindAsync(id);
-            _context.Personos.Remove(person);
+            if (person == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
+            try
+            {
+                _context.Personos.Remove(person);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException /* ex */)
+            {
+                //Log the error (uncomment ex variable name and write a log.)
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
         }
 
         private bool PersonExists(int id)
